Changes: Added Dashing ability Added Prey animations Added another menu theme Reworked player movement Refactored files Updated requirements Updated README Other minor tweaks & additions
