+++ main.py
@@ -1,4 +1,4 @@
-from classes.menu import Menu
+from models.menu import Menu
 
 game_menu = Menu()
 game_menu.launch()
